CA-246770: preemptively catch rtnl_link_alloc_cache allocation errors
The new version of `ocaml-netlink` raises an excepion when
`libnl:rtnl_link_alloc_cache` fails to allocate the link cache with
profiling informations (to prevent a segfault).

We intercept the error when getting link stats and return an empty device
statistics list, this avoid preventing the collection of further bond
performance statistics.

We are not catching `Out_of_memory` errors that could be raised during
the preallocation of the cache memory, nor other errors on purpose:
in these cases the failure should be dealt with by the old mechanism
(`network_monitor_thread.ml:263`).

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>